requirements.txt
flask
playwright
google-genai
python-dotenv
weasyprint
jinja2
requests
Pillow
